feat: add new section after installing the app
